% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{anova_select_options}
\alias{anova_select_options}
\title{Anova Select Options}
\usage{
anova_select_options(fieldbook)
}
\description{
This function give item option for shinySelectize in ANOVA ui
}
\author{
Omar Benites
}

