% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_anova.R
\name{anova_ui}
\alias{anova_ui}
\title{UI for open books in HiDAP
Returns user friendly ui}
\usage{
anova_ui(type = "tab", title = "Analysis of Variance",
  name = "analysis_aov")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for open books in HiDAP
Returns user friendly ui
}
\author{
Omar Benites
}

