% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-prec.R
\name{clean_fb}
\alias{clean_fb}
\title{Fieldbook cleaner for AGROFIMS DATA}
\usage{
clean_fb(fb, trend = FALSE)
}
\arguments{
\item{fb}{data.frame. Field book data}
}
\description{
Clean headers,special characters like "/", whitespaces. In addition, it remove TIMESTAMP columns and empty traits.
}
\author{
Omar Benites
}
