% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_droindex.R
\name{droindex_server}
\alias{droindex_server}
\title{Server for Drought Index}
\usage{
droindex_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for Drought Index
}
\author{
Omar Benites
}
