% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_droindex.R
\name{droindex_ui}
\alias{droindex_ui}
\title{UI for Drought Indexes
Returns user friendly ui}
\usage{
droindex_ui(type = "tab", title = "Drought Index",
  name = "analysis_droindex")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for Drought Indexes
Returns user friendly ui
}
\author{
Omar Benites
}

