% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_transformation.R
\name{dtr_server}
\alias{dtr_server}
\title{Server for Data Transformation}
\usage{
dtr_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for Data Transformation
}
\author{
Omar Benites
}

