% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_transformation.R
\name{dtr_ui}
\alias{dtr_ui}
\title{UI for Data Transformation in HiDAP
Returns user friendly ui}
\usage{
dtr_ui(type = "tab", title = "Data Transformation", name = "analysis_dtr")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for Data Transformation in HiDAP
Returns user friendly ui
}
\author{
Omar Benites
}

