% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_elston.R
\name{elston_server}
\alias{elston_server}
\title{Server Elston Analysis}
\usage{
elston_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server Elston Analysis
}
\author{
Omar Benites
}

