% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_elston_sbase.R
\name{elston_server_sbase}
\alias{elston_server_sbase}
\title{Server Elston Analysis}
\usage{
elston_server_sbase(input, output, session, values)
}
\arguments{
\item{input}{elston shiny server input}

\item{output}{elston shiny server output}

\item{session}{elston shiny server session}

\item{values}{elston reactive values}
}
\description{
Server Elston Analysis
}

