% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_genetic_design.R
\name{genetic_server}
\alias{genetic_server}
\title{Server for Genetic Analysis}
\usage{
genetic_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for Genetic Analysis
}
\author{
Omar Benites
}

