% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_genetic_analysis.R
\name{genetic_ui}
\alias{genetic_ui}
\title{UI for Genetic designs
Returns user friendly ui}
\usage{
genetic_ui(type = "tab", title = "Genetic Analysis",
  name = "genetic_design_analysis")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for Genetic designs
Returns user friendly ui
}
\author{
Omar Benites
}
