% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-agro.R
\name{get_agrofims_file}
\alias{get_agrofims_file}
\title{Get file direcction of agrofim's files.}
\usage{
get_agrofims_file(path = "/home/obenites/AGROFIMS/kdsmart/",
  file_name = "PURI1567089918.xlsx")
}
\arguments{
\item{path}{file path}

\item{file_name}{file name. The file name is provided by AGROFIMS after downloading the fb file. Ex. PURI1567089918}
}
\description{
Get file direcction of agrofim's files.
}
