% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-prec.R
\name{get_design_factors}
\alias{get_design_factors}
\title{Get factors provided by different designs}
\usage{
get_design_factors(design = "Completely Randomized Design (CRD)")
}
\arguments{
\item{fb}{data.frame. Field book data}
}
\description{
character vector. Get factors according to the statistical design
}
\author{
Omar Benites
}
