% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb-prec.R
\name{get_fbsubsample}
\alias{get_fbsubsample}
\title{Get Fieldbook with subsample data according to the traits}
\usage{
get_fbsubsample(fb, design = "Completely Randomized Design (CRD)", trait,
  trend = FALSE, trend_col = NULL)
}
\arguments{
\item{fb}{data.frame. Field book data}

\item{design}{character Whole statistical design name}

\item{trait}{character Trait name in the field book}
}
\description{
Clean headers,
}
\author{
Omar Benites
}
