% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_lda.R
\name{lda_server}
\alias{lda_server}
\title{Server for Linear Discriminant Analysis in HiDAP}
\usage{
lda_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for Linear Discriminant Analysis in HiDAP
}
\author{
Omar Benites
}

