% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_lda.R
\name{lda_ui}
\alias{lda_ui}
\title{UI for Linear Discriminant Analysis in HiDAP}
\usage{
lda_ui(type = "tab", title = "Linear Discriminant Analysis",
  name = "analysis_lda")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
Returns user friendly ui
}
\author{
Omar Benites
}

