% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_met_sbase.R
\name{met_server_sbase}
\alias{met_server_sbase}
\title{Server Multi-Environment Analysis (MET)}
\usage{
met_server_sbase(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server Multi-Environment Analysis (MET)
}
\author{
Omar Benites
}
