% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_met.R
\name{met_ui}
\alias{met_ui}
\title{UI for Multi Environment Analysis (MET) in HiDAP
Returns user friendly ui}
\usage{
met_ui(type = "tab", title = "Multi-Environment", name = "analysis_met")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for Multi Environment Analysis (MET) in HiDAP
Returns user friendly ui
}
\author{
Omar Benites
}

