% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_met_hagrofims.R
\name{met_ui_}
\alias{met_ui_}
\title{UI for Multi Environment Analysis (MET) in HiDAP
Returns user friendly ui}
\usage{
met_ui_(type = "tab",
  title = "The Statistical Analysis of Multiple Statistical",
  name = "analysis_met")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for Multi Environment Analysis (MET) in HiDAP
Returns user friendly ui
}
\author{
Omar Benites
}
