% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_pbaker.R
\name{pbaker_server}
\alias{pbaker_server}
\title{Server pbaker Analysis}
\usage{
pbaker_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server pbaker Analysis
}
\author{
Omar Benites
}

