% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pbaker.R
\name{pbaker_ui}
\alias{pbaker_ui}
\title{UI for pbaker Analysis in HiDAP
Returns user friendly ui}
\usage{
pbaker_ui(type = "tab", title = "Pesek Baker", name = "analysis_pbaker")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for pbaker Analysis in HiDAP
Returns user friendly ui
}
\author{
Omar Benites
}

