% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_pvs_anova.R
\name{pvs_anova_server}
\alias{pvs_anova_server}
\title{Server for PVS' analysis of variance}
\usage{
pvs_anova_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for PVS' analysis of variance
}
\author{
Omar Benites
}
