% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pvs_anova.R
\name{pvs_anova_ui}
\alias{pvs_anova_ui}
\title{UI for PVS's analysis of variance
Returns user friendly ui}
\usage{
pvs_anova_ui(type = "tab", title = "Anova PVS", name = "anova_pvs")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for PVS's analysis of variance
Returns user friendly ui
}
\author{
Omar Benites
}

