% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_pvs.R
\name{pvs_server}
\alias{pvs_server}
\title{Server for PVS report for selection criteria and organoleptics
Returns user friendly ui for selection criteria and organoleptics}
\usage{
pvs_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for PVS report for selection criteria and organoleptics
Returns user friendly ui for selection criteria and organoleptics
}
\author{
Omar Benites
}
