% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_options}
\alias{select_options}
\title{Select Options}
\usage{
select_options(fieldbook)
}
\arguments{
\item{fieldbook}{A data.frame with the fieldbook data}
}
\description{
This function give item option for shinySelectize in ANOVA ui
}
\author{
Omar Benites
}

