% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{single_error}
\alias{single_error}
\title{Error handler in fieldbook analysis}
\usage{
single_error(code, trait)
}
\arguments{
\item{code}{source code to try}

\item{trait}{trait}
}
\description{
Error always happens during data analysis. This function aims to show a error message for end-users to
detect what it wrong in their data.
}
\author{
Omar Benites
}
