% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_single_hdagrofims.R
\name{single_hdagrofims_server}
\alias{single_hdagrofims_server}
\title{Server for Single Environment analysis for HIDAP-AgroFIMS}
\usage{
single_hdagrofims_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for Single Environment analysis for HIDAP-AgroFIMS
}
\author{
Omar Benites
}
