% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_single_hdagrofims.R
\name{single_hdagrofims_ui}
\alias{single_hdagrofims_ui}
\title{UI for Single Environment Analysis for HIDAP-AgroFIMS
Returns user friendly ui}
\usage{
single_hdagrofims_ui(type = "tab", title = "Single Fieldbook Analysis",
  name = "analysis_single")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for Single Environment Analysis for HIDAP-AgroFIMS
Returns user friendly ui
}
\author{
Omar Benites
}
