% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_sbase_single.R
\name{single_sbase_ui}
\alias{single_sbase_ui}
\title{UI for Single Environment Analysis in HiDAP
Returns user friendly ui}
\usage{
single_sbase_ui(type = "tab",
  title = "Statistical Analysis for One Environment Trial",
  name = "analysis_single")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for Single Environment Analysis in HiDAP
Returns user friendly ui
}
\author{
Omar Benites
}
