% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_single.R
\name{single_server}
\alias{single_server}
\title{Server for Single Environment analysis}
\usage{
single_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for Single Environment analysis
}
\author{
Omar Benites
}

