% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_trend_hdagrofims.R
\name{trend_hdagrofims_server}
\alias{trend_hdagrofims_server}
\title{Server for Trend Environment analysis for HIDAP-AgroFIMS}
\usage{
trend_hdagrofims_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for Trend Environment analysis for HIDAP-AgroFIMS
}
\author{
Omar Benites
}
