% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_trend_hdagrofims.R
\name{trend_hdagrofims_ui}
\alias{trend_hdagrofims_ui}
\title{UI for Trend Environment Analysis for HIDAP-AgroFIMS
Returns user friendly ui}
\usage{
trend_hdagrofims_ui(type = "tab", title = "Trend Analysis",
  name = "analysis_trend")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for Trend Environment Analysis for HIDAP-AgroFIMS
Returns user friendly ui
}
\author{
Omar Benites
}
