% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pvs.R
\name{ui_pvs}
\alias{ui_pvs}
\title{UI for PVS report for selection criteria and organoleptics
Returns user friendly ui for selection criteria and organoleptics}
\usage{
ui_pvs(type = "tab", title = "PVS Criteria & Organoleptic",
  name = "analysis_pvs")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for PVS report for selection criteria and organoleptics
Returns user friendly ui for selection criteria and organoleptics
}
\author{
Omar Benites
}

