% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_elston.R
\name{elston_ui}
\alias{elston_ui}
\title{UI for Elston Analysis in HiDAP
Returns user friendly ui}
\usage{
elston_ui(type = "tab", title = "Elston", name = "analysis_elston")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for Elston Analysis in HiDAP
Returns user friendly ui
}
\author{
Omar Benites
}

