% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R, R/trmodf.R, R/ui_elston.R
\name{elston_ui}
\alias{elston_ui}
\title{UI for Elston Analysis in HiDAP
Returns user friendly ui}
\usage{
elston_ui(type = "tab", title = "Elston Index for Ranking and Selection",
  name = "analysis_elston")

elston_ui(type = "tab", title = "Elston Index for Ranking and Selection",
  name = "analysis_elston")

elston_ui(type = "tab", title = "Elston Index for Ranking and Selection",
  name = "analysis_elston")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}

\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}

\item{type}{type of UI element, default is a tab in a shinydashboard for elston index}

\item{title}{display title name for elston index}

\item{name}{UI TabName for elston index}
}
\description{
UI for Elston Analysis in HiDAP
Returns user friendly ui

UI for Elston Analysis in HiDAP
Returns user friendly ui

UI for Elston Analysis in HiDAP
Returns user friendly ui for elston index
}
\author{
Omar Benites

Omar Benites

Omar Benites
}

